Print local timezone offset in activity script
I wanted to be able to copy/paste the output of activity.py to
share with others, and having the RFC3339 timestamp with timezone
helps reduce confusion.

- activity.py
Factored the timestamp formatting code to a single function then
call that everywhere it is needed.

My editor removed some extra space chars on blank lines

- .gitignore
Added .python-version to the ignore list, it's common for folks
using python virtual environments